Oni Pre-Alpha 0.2.7v
The MPS is done, now we are just adding the new types of tiles. We
already start to develop the game core in another branch of the project
but this branch will continue active because i'll start to test the
online integration for the MPS and we'll use it to fix any bug that
happen between the project development.